Feature/inbox

+++ MessageAppRemake/Services/Firebase/UserService.swift
@@ -13,6 +13,7 @@ import FirebaseAuth
 
 protocol UserServiceProtocol {
 	var currentUserPublisher: Published<User?>.Publisher { get }
+	func fetchUser(userId: String)  async throws -> User
 	func fetchCurrentUser() async throws
 	func uploadUserData(user: User, userId: String) async throws
 	func deleteUserData(userId: String) async throws
@@ -32,24 +33,27 @@ final class UserService: UserServiceProtocol {
 	}
 	
 	@MainActor
+	func fetchUser(userId: String) async throws -> User {
+		let document = try await firestore.collection("users").document(userId).getDocument()
+		guard let data = document.data() else {
+			throw AppError(title: "Error", message: "User not found")
+		}
+		return try Firestore.Decoder().decode(User.self, from: data)
+	}
+	
+	@MainActor
 	func fetchCurrentUser() async throws {
 		guard let uid = Auth.auth().currentUser?.uid else {
-			throw AppError(title: "User Not Logged In", message: "Please log in to access your account.")
+			return
 		}
-		
-		let document = try await firestore.collection("users").document(uid).getDocument()
-		guard let data = document.data() else {
-			throw AppError(title: "User Not Found", message: "We couldn't find your account details.")
-		}
-		
-		let user = try Firestore.Decoder().decode(User.self, from: data)
-		self.currentUser = user
+		self.currentUser = try await fetchUser(userId: uid)
 	}
 	
 	func uploadUserData(user: User, userId: String) async throws {
 		do {
 			let encodedUser = try Firestore.Encoder().encode(user)
 			try await firestore.collection("users").document(userId).setData(encodedUser)
+			try await fetchCurrentUser()
 		} catch {
 			throw error
 		}
@@ -63,15 +67,19 @@ final class UserService: UserServiceProtocol {
 		}
 		
 		let profileImageRef = storage.reference().child("ProfileImages").child(userId)
-		do {
-			try await profileImageRef.delete()
-		} catch let error as NSError {
-			if error.domain == StorageErrorDomain && error.code == StorageErrorCode.objectNotFound.rawValue {
-				throw AppError(title: "Delete Failed", message: "No profile image found for deletion.")
-			} else {
-				throw error
+		if let _ = try? await profileImageRef.getMetadata() {
+			do {
+				try await profileImageRef.delete()
+			} catch let error as NSError {
+				if error.domain == StorageErrorDomain && error.code == StorageErrorCode.objectNotFound.rawValue {
+					throw AppError(title: "Delete Failed", message: "No profile image found for deletion.")
+				} else {
+					throw error
+				}
 			}
 		}
+		
+		self.currentUser = nil
 	}
 	
 	func uploadProfileImage(userId: String, imageData: Data) async throws -> String {
